requirements.txt
absl-py==1.4.0
altair==5.0.1
asgiref==3.5.2
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
authentication==1.1.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cryptography==40.0.2
customtkinter==5.1.2
cycler==0.11.0
darkdetect==0.8.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1
django-allauth==0.54.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-froala-editor==4.0.18
django-phonenumber-field==7.0.0
django-rest-framework-nested==0.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.10.7
Flask==2.3.1
flatbuffers==23.3.3
fonttools==4.39.0
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.51.3
grpcio-tools==1.51.3
h11==0.14.0
h5py==3.8.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.21.3
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.0.3
jupyter_core==5.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.1.1
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.17
pydeck==0.8.1b0
Pygments==2.14.0
PyJWT==2.4.0
Pympler==1.0.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pyzmq==25.0.0
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.9.1
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.2
stack-data==0.6.2
streamlit==1.23.1
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tk==0.1.0
tkintertable==1.3.3
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2022.1
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
validators==0.20.0
virtualenv==20.21.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.0.1
zipp==3.15.0
